copyDir函数重载的临时解决方案 调整moveDir接口，支持冲突文件信息将在抛出异常的data属性中以Array<ConflictFiles>形式提供 调整copyDir文件名 处理告警 openmode参数解析修复

+++ interfaces/kits/js/src/mod_fs/class_randomaccessfile/fs_randomaccessfile.cpp
@@ -106,7 +106,7 @@ tuple<bool, void *, size_t, int64_t> ValidReadArg(ArrayBuffer &buffer, const opt
         return { false, nullptr, retLen, offset };
     }
     optional<size_t> lengthOp = nullopt;
-    optional<size_t> offsetOp = nullopt;
+    optional<int64_t> offsetOp = nullopt;
     if (options.has_value()) {
         ReadOptions op = options.value();
         lengthOp = op.length;
@@ -161,7 +161,7 @@ tuple<bool, void *, size_t, int64_t> ValidWriteArg(
     }
 
     optional<size_t> lengthOp = nullopt;
-    optional<size_t> offsetOp = nullopt;
+    optional<int64_t> offsetOp = nullopt;
     if (options.has_value()) {
         WriteOptions op = options.value();
         lengthOp = op.length;
